# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Task Manager - Backend - FastAPI",
    "description": "REST API for an task manager api",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/users/": {
      "get": {
        "tags": ["Users"],
        "summary": "Read Users",
        "description": "Retrieve a list of users with pagination, filtering, and sorting. (Admin/Super role required)",
        "operationId": "read_users_api_v1_users__get",
        "security": [{ "OAuth2PasswordBearer": [] }],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 10,
              "title": "Page Size"
            }
          },
          {
            "name": "full_name",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "string" }, { "type": "null" }],
              "title": "Full Name"
            }
          },
          {
            "name": "role",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                { "$ref": "#/components/schemas/RoleEnum" },
                { "type": "null" }
              ],
              "title": "Role"
            }
          },
          {
            "name": "sort_by",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                { "$ref": "#/components/schemas/UserSortOptions" },
                { "type": "null" }
              ],
              "title": "Sort By"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Page_UserOut_" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Users"],
        "summary": "Create User",
        "description": "Create a new user. (Admin/Super role required)",
        "operationId": "create_user_api_v1_users__post",
        "security": [{ "OAuth2PasswordBearer": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UserCreate" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UserOut" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/api/v1/users/emails": {
      "get": {
        "tags": ["Users"],
        "summary": "Read Users Emails",
        "description": "Retrieve a list of all user emails. (Admin/Super role required)",
        "operationId": "read_users_emails_api_v1_users_emails_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": { "type": "string" },
                  "type": "array",
                  "title": "Response Read Users Emails Api V1 Users Emails Get"
                }
              }
            }
          }
        },
        "security": [{ "OAuth2PasswordBearer": [] }]
      }
    },
    "/api/v1/users/{user_id}": {
      "get": {
        "tags": ["Users"],
        "summary": "Read User by ID",
        "description": "Retrieve a single user by their ID. (Admin/Super role required)",
        "operationId": "read_user_api_v1_users__user_id__get",
        "security": [{ "OAuth2PasswordBearer": [] }],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "title": "User Id" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UserOut" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["Users"],
        "summary": "Update User",
        "description": "Update an existing user by their ID. (Admin/Super role required)",
        "operationId": "update_user_api_v1_users__user_id__put",
        "security": [{ "OAuth2PasswordBearer": [] }],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "title": "User Id" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UserUpdate" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UserOut" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Users"],
        "summary": "Delete User (Deactivate)",
        "description": "Deactivate a user by their ID (soft delete). (Admin/Super role required)",
        "operationId": "delete_user_api_v1_users__user_id__delete",
        "security": [{ "OAuth2PasswordBearer": [] }],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "title": "User Id" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UserOut" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": { "$ref": "#/components/schemas/ValidationError" },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Page_UserOut_": {
        "properties": {
          "items": {
            "items": { "$ref": "#/components/schemas/UserOut" },
            "type": "array",
            "title": "Items"
          },
          "total_items": { "type": "integer", "title": "Total Items" },
          "total_pages": { "type": "integer", "title": "Total Pages" },
          "page": { "type": "integer", "title": "Page" },
          "page_size": { "type": "integer", "title": "Page Size" }
        },
        "type": "object",
        "required": [
          "items",
          "total_items",
          "total_pages",
          "page",
          "page_size"
        ],
        "title": "Page[UserOut]"
      },
...
      "RoleEnum": {
        "type": "string",
        "enum": ["ADMIN", "SUPER", "COMMON"],
        "title": "RoleEnum"
      },
...
      "UserCreate": {
        "properties": {
          "email": { "type": "string", "format": "email", "title": "Email" },
          "full_name": { "type": "string", "title": "Full Name" },
          "role": { "$ref": "#/components/schemas/RoleEnum" },
          "password": { "type": "string", "minLength": 6, "title": "Password" }
        },
        "type": "object",
        "required": ["email", "full_name", "role", "password"],
        "title": "UserCreate"
      },
...
      "UserOut": {
        "properties": {
          "email": { "type": "string", "format": "email", "title": "Email" },
          "full_name": { "type": "string", "title": "Full Name" },
          "role": { "$ref": "#/components/schemas/RoleEnum" },
          "id": { "type": "integer", "title": "Id" },
          "is_active": { "type": "boolean", "title": "Is Active" }
        },
        "type": "object",
        "required": ["email", "full_name", "role", "id", "is_active"],
        "title": "UserOut"
      },
...
      "UserSortOptions": {
        "type": "string",
        "enum": ["full_name_asc", "full_name_desc", "email_asc", "email_desc"],
        "title": "UserSortOptions"
      },
      "UserUpdate": {
        "properties": {
          "full_name": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Full Name"
          },
          "is_active": {
            "anyOf": [{ "type": "boolean" }, { "type": "null" }],
            "title": "Is Active"
          },
          "role": {
            "anyOf": [
              { "$ref": "#/components/schemas/RoleEnum" },
              { "type": "null" }
            ]
          }
        },
        "type": "object",
        "title": "UserUpdate"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": { "anyOf": [{ "type": "string" }, { "type": "integer" }] },
            "type": "array",
            "title": "Location"
          },
          "msg": { "type": "string", "title": "Message" },
          "type": { "type": "string", "title": "Error Type" }
        },
        "type": "object",
        "required": ["loc", "msg", "type"],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": { "password": { "scopes": {}, "tokenUrl": "/auth/login" } }
      }
    }
  }
}